requirements.txt
absl-py
jax
jaxlib
matplotlib
numpy
scipy
torch
torchaudio
